improved tests cases; fixed bug in RemoveDups2
