+++ include/2.1.h
@@ -5,10 +5,8 @@
 #include <string>
 
 namespace CtCI {
-std::string ListToString(std::list<int> L);
 std::list<int> RemoveDups1(std::list<int> L);
 std::list<int> RemoveDups2(std::list<int> L);
 std::list<int> RemoveDups3(std::list<int> L);
-
 }  // namespace CtCI
 #endif  // _2_1_H
